renamed ssr.py to regressor_sequence.py
