+++ pysr/regressor_sequence.py
